+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### What's Changed
 
 * Revert nodes get functions to default components to `NodeComponents.NONE` instead of `ALL`.
+* Added `Enlkind` endpoint and models.
 * Update testing/linting configuration.
 
 **Full Changelog**: https://github.com/mmahacek/PyONMS/compare/v0.1.2...v0.1.3